+++ src/app/edit-user/edit-user.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EditUserComponent } from './edit-user.component';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ButtonsModule, BsModalRef, BsModalService, ModalModule } from 'ngx-bootstrap';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { By } from '@angular/platform-browser';
@@ -20,6 +20,7 @@ describe('EditUserComponent', () => {
         ButtonsModule,
         RouterTestingModule,
         ModalModule.forRoot(),
+        ReactiveFormsModule
       ],
       providers: [
         BsModalRef,