+++ src/app/user-info/user-info.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { UserInfoComponent } from './user-info.component';
 import { ButtonsModule, BsModalRef, BsModalService, ModalModule } from 'ngx-bootstrap';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('UserInfoComponent', () => {
@@ -17,6 +17,7 @@ describe('UserInfoComponent', () => {
         FormsModule,
         ButtonsModule,
         ModalModule.forRoot(),
+        ReactiveFormsModule
       ],
       providers: [
         BsModalRef,
@@ -30,15 +31,6 @@ describe('UserInfoComponent', () => {
     fixture = TestBed.createComponent(UserInfoComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
-  });
-
-  it('checkUserRole() - Should check the role of current user', () => {
-    const checkUserRole = spyOn(component, 'checkUserRole').and.returnValue(true);
-    const buttonRole = {};
-
-    checkUserRole(buttonRole);
-
-    expect(checkUserRole).toHaveBeenCalledWith(buttonRole);
   });
 
   it('deleteUser() - Should test if the method is called with parameter as mock user object', () => {